stagionalita/requirements.txt
streamlit>=1.37.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
plotly>=5.18.0
requests>=2.31.0
Pillow>=10.0.0
schedule>=1.2.0
